Revert "Graphql queries (#68)"
This reverts commit 5982ee773092935cc09188867e718607d6d93644.
