--- client/@types/is-hotkey/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'is-hotkey'
