--- client/@types/slate-react/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'slate-react'
