--- client/@types/slate/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'slate'
