+++ docker-compose.yml
@@ -1,0 +1,38 @@
+version: '3'
+
+services:
+  tsseract-app:
+    container_name: tsseract-app
+    restart: always
+    build:
+      context: .
+      dockerfile: Dockerfile
+    env_file: .env
+    depends_on:
+      - tsseract-db
+    expose:
+      - '3000'
+      - '8080'
+    ports:
+      - '3000:3000'
+      - '8080:8080'
+    links:
+      - tsseract-db
+    volumes:
+      - ${PWD}/:/tsseract
+    stdin_open: true
+    networks:
+      - tsseract-network
+
+  tsseract-db:
+    image: mongo
+    container_name: tsseract-db
+    restart: always
+    ports:
+      - '27017:27017'
+    networks:
+      - tsseract-network
+
+networks:
+  tsseract-network:
+    driver: bridge
