+++ package.json
@@ -32,8 +32,6 @@
     "dayjs": "^1.10.4",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "express-graphql": "^0.12.0",
-    "graphql": "^15.5.0",
     "is-hotkey": "^0.1.6",
     "is-url": "^1.2.4",
     "jsonwebtoken": "^8.5.1",
@@ -59,8 +57,6 @@
     "@types/node": "^14.0.22",
     "@types/react": "^16.9.43",
     "@types/set-cookie-parser": "0.0.6",
-    "@types/slate": "^0.47.8",
-    "@types/slate-react": "^0.22.9",
     "@types/supertest": "^2.0.10",
     "babel-plugin-styled-components": "^1.11.1",
     "faker": "^5.1.0",
