+++ server/models/user.ts
@@ -1,7 +1,7 @@
-import Joi from '@hapi/joi';
 import { Schema, Types, model } from 'mongoose';
+import Joi from '@hapi/joi';
 
-import regex from '../../helpers/regex';
+import regex from '../helpers/regex';
 
 const userSchema = new Schema({
   name: {
@@ -58,7 +58,7 @@ const userSchema = new Schema({
   },
 });
 
-const Users = model('Users', userSchema);
+export default model('Users', userSchema);
 
 export const validateUser = <T>(user: T) => {
   const schema = Joi.object({
@@ -84,5 +84,3 @@ export const validateUser = <T>(user: T) => {
 
   return schema.validate(user);
 };
-
-export default Users;
