+++ server/tsconfig.json
@@ -16,6 +16,5 @@
     "resolveJsonModule": true,
     "noEmit": false,
     "typeRoots": ["@types", "../node_modules/@types"]
-  },
-  "exclude": ["node_modules"]
+  }
 }
